requirements.txt
git+git://projects.archlinux.org/users/remy/pyalpm.git#egg=pyalpm
PyYAML
docopt
clint